Depend on tokenbridge contracts (#31)
This PR is pulling in the [xDAI tokenbridge repo](https://github.com/poanetwork/tokenbridge-contracts) as a submodule. This is needed so that we can build a custom version of xOWL that is based on their standard bridge contract (`PermittableToken.sol`). They currently don't expose their contracts as an npm module and thus using a submodule here seems most pragmatic to me.

Also their contracts require solidity 0.4.24 which is why I'm adding the relevant config changes in this PR (so that the next PR can focus on the smart contract code + test)

Happy to hear other suggestions.

### Test Plan

Making sure we can lint and compile everything without errors.