Implement Bridge Token OWL (#32)
Implement a custom burnOWL method for our xDAI deployment of xOWL. For now the logic just credits the burnt OWL to a designated receiver. The idea is that the receiver will send the burnt OWL to mainnet where they can be burnt for real (adjusting totalSupply). 
For now, this receiver can be an EOA or multisig under our control - eventually it could become a smart contract that can be permissionlessly invoked to `transferAndCall` burn on mainnet.

I'm thinking about putting a proxy in front of this implementation so that we could even change the logic altogether later on. What do you guys think about that?

### Test Plan

Unit tests